+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 === 0.5.0.beta
 * Model.new with invalid parameter will not raise exception by default now.
 * "id" attribute always being added to model by default now.
+* navigation: true by default for screen initialization.
 
 === 0.4.4
 * Added Prime.env support